Release referentially transparent Metadata module separately (closes #96)
