+++ modules/core/src/main/scala/com.snowplowanalytics.snowplow/scalatracker/Emitter.scala
@@ -12,6 +12,8 @@
  */
 package com.snowplowanalytics.snowplow.scalatracker
 
+import java.util.UUID
+
 /**
  * Emitters are entities in charge of transforming events sent from tracker
  * into actual HTTP requests (IO), which includes:
@@ -19,16 +21,32 @@ package com.snowplowanalytics.snowplow.scalatracker
  * + Queuing `EmitterPayload`
  * + Transforming `EmitterPayload` into Bytes
  * + Backup queue and callbacks
+ *
+ * This trait also includes two methods unrelated to actual sending: `getCurrentMilliseconds` and `generateUUID`.
+ * They are extracted here because `Tracker` is referentially transparent, and we leave it up for the emitters
+ * to decide whether `F` will be referentially transparent also.
  */
 trait Emitter[F[_]] {
   import Emitter._
 
   /**
-   * Method called to send an event from the tracker to the emitter
+   * Emits the event payload
    *
    * @param event Fully assembled event
    */
   def send(event: EmitterPayload): F[Unit]
+
+  /**
+   * Returns the current time in milliseconds. Most likely it will be `System.currentTimeMillis()` wrapped in `F`
+   * @return current time in milliseconds
+   */
+  def getCurrentMilliseconds: F[Long]
+
+  /**
+   * Returns a random UUID. Most likely it will be `UUID.randomUUID()` wrapped in `F`
+   * @return current time in milliseconds
+   */
+  def generateUUID: F[UUID]
 }
 
 object Emitter {
