+++ modules/id-emitter/src/main/scala/com/snowplowanalytics/snowplow/scalatracker/emitters/id/BaseEmitter.scala
@@ -1,0 +1,14 @@
+package com.snowplowanalytics.snowplow.scalatracker.emitters.id
+
+import java.util.UUID
+
+import cats.Id
+import com.snowplowanalytics.snowplow.scalatracker.Emitter
+
+abstract class BaseEmitter extends Emitter[Id] {
+
+  override def generateUUID: Id[UUID] = UUID.randomUUID()
+
+  override def getCurrentMilliseconds: Id[Long] = System.currentTimeMillis()
+
+}
